+++ hrosailing/core/__init__.py
@@ -1,5 +1,13 @@
-import .modelfunctions
-import .computing
-import .constants
-import .data
-import .statistics
+import hrosailing.core.modelfunctions
+import hrosailing.core.computing
+import hrosailing.core.constants
+import hrosailing.core.data
+import hrosailing.core.statistics
+
+__all__ = [
+    "modelfunctions",
+    "statistics",
+    "computing",
+    "data",
+    "constants"
+]
