+++ hrosailing/pipeline/__init__.py
@@ -11,3 +11,10 @@ from .extensions import (
 from .quality_assurance import (
     QualityAssurance, ComformingQualityAssurance, MinimalQualityAssurance
 )
+
+__all__ = [
+    "PipelineOutput", "PolarPipeline",
+    "Expander", "LazyExpander", "WeatherExpander",
+    "PipelineExtension", "TableExtension", "PointcloudExtension", "CurveExtension",
+    "QualityAssurance", "ComformingQualityAssurance", "MinimalQualityAssurance"
+]
